🎨 데이터 흐름 수정
