+++ app/types/index.ts
@@ -1,0 +1,7 @@
+export interface Post {
+  id: string;
+  title: string;
+  parent_id: string;
+  type: string;
+  sub_blog: string;
+}
